\name{tufteBarTheme}
\alias{tufteBarTheme}

\title{
Add Tufte's style to a ggplot2 bar plot.
}
\description{
This mainly serves as support function for tufteBar().
}
\usage{
tufteTheme(tuftePlot, x.text.rotate, y.text.rotate, font, font_size, x.text.color, y.text.color)
}

\arguments{
  \item{tuftePlot}{

}
  \item{font}{

}
  \item{font_size}{

}
}
\details{

}
\value{

}
\references{

}
\author{

}
\note{

}



\seealso{

}
\examples{

}


\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
